requirements.txt
fastapi
uvicorn
pymongo
pydantic[dotenv]
PyJWT
python-jose